@@ -24,11 +24,13 @@ def read_text(
     if delimiter is None:
         raise ValueError("delimiter needs to be provided")
 
-    filepath_or_buffer, _ = ioutils.get_reader_filepath_or_buffer(
+    filepath_or_buffer = ioutils.get_reader_filepath_or_buffer(
         path_or_data=filepath_or_buffer,
-        compression=None,
         iotypes=(BytesIO, StringIO),
         storage_options=storage_options,
+    )
+    filepath_or_buffer = ioutils._select_single_source(
+        filepath_or_buffer, "read_text"
     )
 
     return cudf.Series._from_data(
